crank/requirements.txt
solana-py>=0.34.0
solders>=0.21.0
httpx>=0.27.0
python-dotenv>=1.0.0
